death, win condition
Now when a player dies they return to their spawn location with full health.

Their spawn location is just whereever they were when the game started

When a player reaches 5 kills the gameController recognizes that player as the winner. So far that means nothing besides a Debug.Log()

Also audio should be fixed